pyproject.toml
[tool.poetry]
name = "packs"
version = "1.0.0"
description = "Some of the packs on https://vanillatweaks.net"
authors = ["VanillaTweaks <team@vanillatweaks.net>"]

[tool.poetry.dependencies]
python = "^3.10"
beet = ">=0.64.0"
mecha = ">=0.48.3"
bolt = ">=0.3.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
